+++ single_offer/src/testutils.rs
@@ -2,7 +2,7 @@
 
 use crate::{token::Identifier, Price, SingleOfferClient};
 
-use soroban_sdk::{AccountId, BigInt, BytesN, Env};
+use soroban_sdk::{AccountId, BytesN, Env};
 
 pub fn register_test_contract(e: &Env) -> BytesN<32> {
     e.register_contract(None, crate::SingleOffer {})
@@ -36,11 +36,11 @@ impl SingleOffer {
         self.client().initialize(&admin, token_a, token_b, &n, &d);
     }
 
-    pub fn trade(&self, to: &Identifier, min: &BigInt) {
+    pub fn trade(&self, to: &Identifier, min: &i128) {
         self.client().trade(&to, &min)
     }
 
-    pub fn withdraw(&self, admin: &AccountId, amount: &BigInt) {
+    pub fn withdraw(&self, admin: &AccountId, amount: &i128) {
         self.client().with_source_account(&admin).withdraw(&amount)
     }
